update tensorflow version for security issue
GHSA-844w-j86r-4x2j
low severity
Vulnerable versions: < 1.15
Patched version: 1.15
Impact
A heap buffer overflow in UnsortedSegmentSum can be produced when the Index template argument is int32.
In this case data_size and num_segments fields are truncated from int64 to int32 and can produce negative numbers, resulting in accessing out of bounds heap memory.

This is unlikely to be exploitable and was detected and fixed internally.
We are making the security advisory only to notify users that it is better to update to TensorFlow 1.15 or 2.0 or later as these versions already have this fixed.

Patches
Patched by db4f9717c41bccc3ce10099ab61996b246099892 and released in all official releases after 1.15 and 2.0.

For more information
Please consult SECURITY.md for more information regarding the security model and how to contact us with issues and questions.